Sort output of `docker ps --filter` with order by creation time
This fix tries to address the issue raised in 25374 where the
output of `docker ps --filter` is in random order and
not deterministic.

This fix sorts the list of containers by creation time so that the
output is deterministic.

An integration test has been added.

This fix fixes 25374.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>
(cherry picked from commit 685307b86a783e06b7086b8dd30070786a9c9eeb)
Signed-off-by: Tibor Vass <tibor@docker.com>