Fix `service update --env-add` issue
This fix tries to address the issue in 25404 where updating environmental
variable in `service update --env-add` will not work.

The issue is because `--env-add` will only append the env, not update if
the same env already exist.

This fix tracks the env variable with a map and update if the variable
is the same.

An integration test has been added.

This fix fixes 25404.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>
(cherry picked from commit b16a22a16e0773f43a7ffba20fe4e2bb36432764)
Signed-off-by: Tibor Vass <tibor@docker.com>